+++ apt/methods/ftp.h
@@ -14,7 +14,7 @@
 class FTPConn
 {
    char Buffer[1024*10];
-   unsigned long Len;
+   unsigned long long Len;
    std::unique_ptr<MethodFd> ServerFd;
    int DataFd;
    int DataListenFd;
@@ -54,9 +54,9 @@ class FTPConn
    bool ExtGoPasv();
    
    // Query
-   bool Size(const char *Path,unsigned long &Size);
+   bool Size(const char *Path,unsigned long long &Size);
    bool ModTime(const char *Path, time_t &Time);
-   bool Get(const char *Path,FileFd &To,unsigned long Resume,
+   bool Get(const char *Path,FileFd &To,unsigned long long Resume,
 	    Hashes &MD5,bool &Missing);
    
    FTPConn(URI Srv);
